crawlers/BaoDienTu/vnexpress/requirements.txt
pymongo==3.11.0
PyYAML==5.3.1
requests==2.24.0
beautifulsoup4==4.9.2
python-engineio==3.14.2
python-socketio==4.6.0
lxml==4.6.1